requirements.txt
bokeh
jupyter-book
mdformat
mdformat-black
mdformat-myst
networkx
sphinx-proof
sphinx-thebe
geopandas
matplotlib
-e .
